update from review comments
