+++ src/Menus.js
@@ -10,8 +10,7 @@ define(function (require, exports, module) {
     
     // Load dependent modules
     var Commands                = require("Commands"),
-        CommandManager          = require("CommandManager"),
-        PerfUtils               = require("PerfUtils");
+        CommandManager          = require("CommandManager");
     
     function init() {
         // Implements the File menu items
@@ -31,21 +30,8 @@ define(function (require, exports, module) {
             CommandManager.execute(Commands.FILE_QUIT);
         });
 
-        // Implements the 'Run Tests' menu to bring up the Jasmine unit test window
-        var testWindow = null;
         $("#menu-debug-runtests").click(function () {
-            if (testWindow) {
-                try {
-                    testWindow.location.reload();
-                } catch (e) {
-                    testWindow = null;  // the window was probably closed
-                }
-            }
-
-            if (!testWindow) {
-                testWindow = window.open("../test/SpecRunner.html");
-                testWindow.location.reload(); // if it was opened before, we need to reload because it will be cached
-            }
+            CommandManager.execute(Commands.DEBUG_RUN_UNIT_TESTS);
         });
         
         // Other debug menu items
@@ -58,56 +44,7 @@ define(function (require, exports, module) {
         });
         
         $("#menu-debug-show-perf").click(function () {
-            var perfHeader = $("<div class='modal-header' />")
-                .append("<a href='#' class='close'>&times;</a>")
-                .append("<h3 class='dialog-title'>Performance Data</h3>");
-            
-            var perfBody = $("<div class='modal-body' style='padding: 0' />");
-
-            var data = $("<table class='zebra-striped condensed-table' style='max-height: 600px; overflow: auto;'>")
-                .append("<thead><th>Operation</th><th>Time (ms)</th></thead>")
-                .append("<tbody />")
-                .appendTo(perfBody);
-            
-            var makeCell = function (content) {
-                return $("<td/>").text(content);
-            };
-            
-            var getValue = function (entry) {
-                // entry is either an Array or a number
-                // If it is an Array, return the average value
-                if (Array.isArray(entry)) {
-                    var i, sum = 0;
-                    
-                    for (i = 0; i < entry.length; i++) {
-                        sum += entry[i];
-                    }
-                    return String(Math.floor(sum / entry.length)) + " (avg)";
-                } else {
-                    return entry;
-                }
-            };
-                
-            var testName;
-            var perfData = PerfUtils.perfData;
-            for (testName in perfData) {
-                if (perfData.hasOwnProperty(testName)) {
-                    // Add row to error table
-                    var row = $("<tr/>")
-                        .append(makeCell(testName))
-                        .append(makeCell(getValue(perfData[testName])))
-                        .appendTo(data);
-                }
-            }
-                                                         
-            var perfDlog = $("<div class='modal hide' />")
-                .append(perfHeader)
-                .append(perfBody)
-                .appendTo(document.body)
-                .modal({
-                    backdrop: "static",
-                    show: true
-                });
+            CommandManager.execute(Commands.DEBUG_SHOW_PERF_DATA);
         });
     }
